[refactor] - HomeViewModel 함수 분리
