+++ presentation/component/src/main/java/com/umc/component/component/UText.kt
@@ -9,12 +9,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextLayoutResult
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.Density
+import com.umc.component.theme.neutral800
+import com.umc.component.theme.primary600
 
 @Composable
 fun UText(
@@ -85,5 +90,22 @@ fun HuggText(
             minLines = minLines,
             onTextLayout = onTextLayout ?: {}
         )
+    }
+}
+
+/**
+ * 문장에서 특정 부분을 강조할 때 사용하는 텍스트 포맷
+ * -> Home 화면의 프로필 카드에서 'X일째 성장하고 있어요' 텍스트 전용
+ * 
+ * **/
+@Composable
+fun getGrowthText(day: Int): AnnotatedString = buildAnnotatedString {
+    // 강조할 부분 (색상 적용)
+    withStyle(style = SpanStyle(color = primary600())) {
+        append("${day}일째")
+    }
+    // 나머지 부분
+    withStyle(style = SpanStyle(color = neutral800())) {
+        append(" 성장하고 있어요")
     }
 }
