+++ presentation/component/src/main/java/com/umc/component/theme/AppStrings.kt
@@ -280,6 +280,7 @@ object AppStrings {
     const val REWARD_ETC_CONTENT = "기타 상벌점 사유와 점수를 입력해 주세요."
     const val REWARD = "상점"
     const val PUNISH = "벌점"
+    const val HAB = "총합"
     const val REWARD_ETC_REASON = "사유"
     const val REWARD_TITLE = "상점 부여하기"
     const val REWARD_ETC_PLACEHOLDER = "사유를 입력해주세요"
