+++ presentation/home/src/main/java/com/umc/presentation/home/schedule/detail/ScheduleDetailScreen.kt
@@ -32,6 +32,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.umc.component.R
 import com.umc.component.component.UBasicDialog
 import com.umc.component.component.UButton
@@ -58,7 +59,7 @@ fun ScheduleDetailRoute(
     viewModel : ScheduleDetailViewModel = hiltViewModel()
 ){
 
-    val uiState by viewModel.uiState.collectAsState()
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
 
     //뒤로 가기 디스패처
     val onBackPressedDispatcher = LocalOnBackPressedDispatcherOwner.current?.onBackPressedDispatcher
