+++ presentation/home/src/main/java/com/umc/presentation/home/schedule/dialog/ScheduleCategoryDialog.kt
@@ -27,6 +27,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.umc.component.component.UButton
 import com.umc.component.component.UText
 import com.umc.component.theme.AppStrings
@@ -46,7 +47,7 @@ fun ScheduleCategoryBottomSheet(
     onDismissRequest: () -> Unit, //종료될때
     onConfirm: () -> Unit //확인 누를 시
 ) {
-    val uiState by viewModel.uiState.collectAsState()
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
     val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
 
     ModalBottomSheet(
